WIP: structuredByLanguage flag in thesaurus instead of thesaurusSet.
Controling the structure by language at the thesaurus level makes more
sense. It is the thesaurus the one that is structured. Thus, this makes
it more modular.
For the flags applyToColNames and applyToColValues, the situation is not
the same. They control how each thesaurus is used when applying the
thesaurus set, not the structure or nature of each thesaurus itself.